feat: native macOS Apple Silicon support with a VideoToolbox media provider (#294)
## Summary

Native macOS (Apple Silicon) support for Bloom: platform providers, a
VideoToolbox/AVFoundation media provider, export wiring, tests, CI, and
packaging.

## What changed

- **Platform**: POSIX/macOS staged-artifact publication
(`staged_artifact_macos`), a `posix_spawn` process supervisor
(`process_supervisor_macos`), and a CoreText font catalogue
(`font_catalog_macos`); Objective-C++ is enabled where Apple frameworks
are used.
- **Media**: the isolated worker's Apple-native provider implements
probe, H.264/HEVC/ProRes decode to YUV420P, AAC/PCM audio decode, and
ProRes/H.264 encode with PCM/AAC audio through AVAssetWriter, with
reopen verification. The host selects the provider per platform and the
worker manifest declares only implemented roles.
- **Export**: composition export, output analysis, and the
frame-preservation report select the platform provider; correct Apple
ProRes provenance wording on macOS; platform-aware preset availability
(ProRes/H.264 offered, FFmpeg-only presets reported unavailable).
- **UI**: bundled-font role resolution tolerates platform family
normalization; availability-aware export defaults; explicit QSettings
flushing; a non-reentrant export poll.
- **Tests**: end-to-end provider, export (through the real worker
process), and decode tests on macOS; platform-aware assertions
elsewhere.
- **CI/packaging**: non-required macOS native and Qt desktop jobs, and
an optional `macdeployqt` target that bundles Qt frameworks into
`Bloom.app`.

## Validation

- macOS: full native and desktop builds; VideoToolbox provider, export,
and decode tests pass; ProRes and H.264 import and export verified in
the app.
- Linux: semantics are unchanged (platform branches and platform-scoped
test assertions); the Linux Clang quality job remains the automated
merge gate.

## Notes

- The macOS CI jobs are non-required initially; Linux remains the only
required check.

+++ src/media/audio/playback/audio_engine.cpp
@@ -326,8 +326,9 @@ RationalTime AudioEngine::positionNow() const noexcept {
     std::scoped_lock lock(controlMutex_);
     const auto frames = framesWritten_.load(std::memory_order_acquire);
     if (playbackRateNumerator_ == playbackRateDenominator_) {
-        const auto numerator = static_cast<long double>(clockOrigin_.numerator()) * config_.rate +
-                               static_cast<long double>(frames) * clockOrigin_.denominator();
+        const auto numerator =
+            static_cast<long double>(clockOrigin_.numerator()) * config_.rate +
+            static_cast<long double>(frames) * static_cast<long double>(clockOrigin_.denominator());
         const auto denominator =
             static_cast<long double>(clockOrigin_.denominator()) * config_.rate;
         if (numerator >= static_cast<long double>(std::numeric_limits<std::int64_t>::min()) &&
